Refactor the routes
