--- apps/api/src/routes/health.ts
@@ -1,9 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.get("/", function (_request, response) {
-  response.json({ status: "ok" });
-});
-
-export default router;
