+++ apps/api/src/routes/index.ts
@@ -1,15 +1,15 @@
 import express from "express";
 
 // Routes
-import defaultRouter from "./default.js";
-import scenarioRoutes from "./scenarios.js";
-import vatsimRoutes from "./vatsim.js";
+import defaultRoutes from "./default.js";
+import scenarioRoutes from "./scenarios/index.js";
+import vatsimRoutes from "./vatsim/index.js";
 /**
  * Registers all application routes with the Express application.
  * @param app Express application instance
  */
 export default function addRoutes(app: express.Express) {
-  app.use(defaultRouter);
+  app.use(defaultRoutes);
   app.use("/scenarios", scenarioRoutes);
   app.use("/vatsim", vatsimRoutes);
 }
