+++ apps/api/src/routes/scenarios/summary/index.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import getMethod from "./get.js";
+
+const router = Router();
+
+router.use("/", getMethod);
+
+export default router;
